Split scanner suites
